Implement assignment 1
