+++ 1/src/Main.java
@@ -1,10 +1,8 @@
-import java.util.*;
-import models.*;
 import players.*;
 
 public class Main{
   public static void main(String[] args){
-    var player = new Mp3Player(Getpa.getMusicLibrary());
+    var player = new IPlayerImpl(Getpa.getContentLibrary());
     Getpa.doSomething(player);
   }
 }
